+++ DTCD-Visualization-Table/src/Plugin.js
@@ -10,8 +10,11 @@ import {
 
 export class VisualizationText extends PanelPlugin {
 
-  #dataSource;
+  #dataSourceName;
   #storageSystem;
+  #guid;
+  #eventSystem;
+  #dataSourceSystemGUID;
 
   static getRegistrationMeta() {
     return pluginMeta;
@@ -24,7 +27,10 @@ export class VisualizationText extends PanelPlugin {
     const eventSystem = new EventSystemAdapter(guid);
 
     eventSystem.registerPluginInstance(this);
+    this.#guid = guid;
+    this.#eventSystem = eventSystem;
     this.#storageSystem = new StorageSystemAdapter();
+    this.#dataSourceSystemGUID = this.getGUID(this.getSystem('DataSourceSystem'));
 
     const { default: VueJS } = this.getDependence('Vue');
 
@@ -34,16 +40,35 @@ export class VisualizationText extends PanelPlugin {
     }).$mount(selector);
 
     this.vueComponent = view.$children[0];
-    this.#dataSource = '';
+    this.#dataSourceName = '';
   }
 
   setPluginConfig(config = {}) {
     const { dataSource } = config;
     if (typeof dataSource !== 'undefined') {
-      this.#dataSource = dataSource;
-      const DS = this.getSystem('DataSourceSystem').getDataSource(this.#dataSource);
+      if (this.#dataSourceName) {
+        this.#eventSystem.unsubscribe(
+          this.#dataSourceSystemGUID,
+          'DataSourceStatusUpdate',
+          this.#guid,
+          'processDataSourceEvent',
+          { dataSource: this.#dataSourceName, status: 'success' },
+        );
+      }
+
+      this.#dataSourceName = dataSource;
+
+      this.#eventSystem.subscribe(
+        this.#dataSourceSystemGUID,
+        'DataSourceStatusUpdate',
+        this.#guid,
+        'processDataSourceEvent',
+        { dataSource, status: 'success' }
+      );
+
+      const DS = this.getSystem('DataSourceSystem').getDataSource(this.#dataSourceName);
       if (DS.status === 'success') {
-        const data = this.#storageSystem.session.getRecord(this.#dataSource);
+        const data = this.#storageSystem.session.getRecord(this.#dataSourceName);
         this.loadData(data);
       }
     }
@@ -51,7 +76,7 @@ export class VisualizationText extends PanelPlugin {
 
   getPluginConfig() {
     const config = {};
-    if (this.#dataSource) config.dataSource = this.#dataSource;
+    if (this.#dataSourceName) config.dataSource = this.#dataSourceName;
     return config;
   }
 
@@ -62,8 +87,8 @@ export class VisualizationText extends PanelPlugin {
 
   processDataSourceEvent(eventData) {
     const { dataSource, status } = eventData;
-    this.#dataSource = dataSource;
-    const data = this.#storageSystem.session.getRecord(this.#dataSource);
+    this.#dataSourceName = dataSource;
+    const data = this.#storageSystem.session.getRecord(this.#dataSourceName);
     this.loadData(data);
   }